Add files via upload
A small program to read the log that I'm seeing in the project. ( well the vendor is just too bad so I have to do it by myself even I'm not really a programmer )
It has certain format so basically I want to know what invoices and PO has been sent to the robot for running.