Update 002B_Read_a_log.py
After the system increase its working threads.. I have to use the transaction id as the unique identifier ( which i should but it will be more complicated )
Also the log is getting huge, the re.research function is consuming unacceptable time to complete. ( for every single ID )

So I split the file by two key words "Received" and "Send". In Received I extract all the ID ( here is all the data send to the system ), and in Send ( here is system back feed ), I use the id to look up the ending time.

I put the data from Send into a dictionary which solved dramatically reduce the running time.